day7/bluetooth-on-esp/hand-bt/pyproject.toml
[project]
name = "hand-bt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mediapipe>=0.10.21",
    "opencv-python>=4.11.0.86",
    "pyserial>=3.5",
]
